modify

+++ benchmarks/benchmark_triton_impl/ops/fwd_kernel_v0.py
@@ -24,7 +24,7 @@ def fwd_kernel_v0(
 
     print(f"Q: {Q}, K: {K}, V: {V}")
 
-    q_head_val = tl.load(Q + tl.arange(0, 256))
+    q_head_val = tl.load(Q + tl.arange(0, 256)).to(tl.float32)
     print(f"q_head_val={q_head_val}")
 
     ##### get offset
